requirements.txt
Django==1.6.2
Pillow==2.3.1
PyYAML==3.11
South==0.8.2
amqp==1.4.5
anyjson==0.3.3
argh==0.24.1
billiard==3.3.0.17
-e git+https://github.com/boto/boto.git@8e832923a642115dbcae8cb103898d351d3bdaca#egg=boto-dev
celery==3.1.12
django-cacheops==2.1.1
django-debug-toolbar==1.2
django-redis==3.6.2
django-ses==0.6.0
django-storages==1.1.8
django-tastypie==0.11.0
-e git+https://github.com/pythonforfacebook/facebook-sdk.git@e65d06158e48388b3932563f1483ca77065951b3#egg=facebook_sdk-dev
funcy==1.1
httplib2==0.8
jsonfield==0.9.20
kombu==3.0.19
meld3==1.0.0
mimeparse==0.1.3
oauth2==1.5.211
oauthlib==0.6.0
pathtools==0.1.2
psycopg2==2.5.2
pycrypto==2.6.1
python-dateutil==2.2
python-memcached==1.53
python-mimeparse==0.1.4
python-oauth2==0.6.0
python-openid==2.2.5
python-social-auth==0.1.22
pytz==2013b
redis==2.9.1
requests==2.0.1
requests-oauthlib==0.4.0
simplejson==3.4.1
six==1.4.1
-e git+https://github.com/mariocesar/sorl-thumbnail.git@18ea5230d9a98fdf674af8c9eea2c70f101fb454#egg=sorl_thumbnail-dev
sqlparse==0.1.11
supervisor==3.0
watchdog==0.7.1
wsgiref==0.1.2
